% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp.R
\docType{data}
\name{allmeta}
\alias{allmeta}
\title{1000 Genomes, SGDP, HGDP, and GGVP metadata}
\format{
A tibble with 212 rows and 8 columns fixme:
\describe{
  \item{pop}{Short population code}
  \item{reg}{Short region code}
  \item{population}{Long population description}
  \item{region}{Long region description}
  \item{regcolor}{Color for plotting this region on a map}
  \item{lat}{Population latitude}
  \item{lng}{Population longitude}
}
}
\source{
\url{https://www.internationalgenome.org/api/beta/population/_search/igsr_populations.tsv}
}
\usage{
allmeta
}
\description{
Fixme description
}
\references{
Byrska-Bishop, Marta, et al. "High-coverage whole-genome sequencing of the expanded 1000 Genomes Project cohort including 602 trios." Cell 185.18 (2022): 3426-3440.

1000 Genomes Project Consortium. "A global reference for human genetic variation." Nature 526.7571 (2015): 68.

License information is available at \url{https://github.com/igsr/1000Genomes_data_indexes/blob/master/LICENSE}.
}
\keyword{datasets}
