% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp-package.R
\docType{package}
\name{kgp-package}
\alias{kgp}
\alias{kgp-package}
\title{kgp: 1000 Genomes Project Metadata}
\description{
This data package provides metadata about populations and data about samples from the 1000 Genomes Project, including the 2,504 samples sequenced for the Phase 3 release and the expanded collection of 3,202 samples with 602 additional trios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/kgp}
  \item \url{https://stephenturner.github.io/kgp/}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
