% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp.R
\docType{data}
\name{kgp3}
\alias{kgp3}
\title{1000 Genomes Project sample data (Phase 3)}
\format{
A tibble with 2504 rows and 10 columns:
\describe{
  \item{fid}{Family ID}
  \item{id}{Individual ID}
  \item{pid}{Paternal ID}
  \item{mid}{Maternal ID}
  \item{sex}{Sex (1=Male, 2=Female)}
  \item{sexf}{Sex as a factor}
  \item{pop}{Short population code}
  \item{reg}{Short region code}
  \item{population}{Long population description}
  \item{region}{Long region description}
}
}
\source{
\url{http://ftp.1000genomes.ebi.ac.uk/vol1/ftp/}
}
\usage{
kgp3
}
\description{
Sample, pedigree, and population data for 2,504 samples in the Phase 3 release of the 1000 Genomes Project data.
}
\references{
Byrska-Bishop, Marta, et al. "High-coverage whole-genome sequencing of the expanded 1000 Genomes Project cohort including 602 trios." Cell 185.18 (2022): 3426-3440.

1000 Genomes Project Consortium. "A global reference for human genetic variation." Nature 526.7571 (2015): 68.

License information is available at \url{https://github.com/igsr/1000Genomes_data_indexes/blob/master/LICENSE}.
}
\keyword{datasets}
