% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp.R
\docType{data}
\name{kgpe}
\alias{kgpe}
\title{1000 Genomes Project sample data (Expanded)}
\format{
A tibble with 3202 rows and 11 columns:
\describe{
  \item{fid}{Family ID}
  \item{id}{Individual ID}
  \item{pid}{Paternal ID}
  \item{mid}{Maternal ID}
  \item{sex}{Sex (1=Male, 2=Female)}
  \item{sexf}{Sex as a factor}
  \item{pop}{Short population code}
  \item{reg}{Short region code}
  \item{population}{Long population description}
  \item{region}{Long region description}
  \item{phase3}{Logical; indicates whether this sample is included in the Phase 3 release data}
}
}
\source{
\url{http://ftp.1000genomes.ebi.ac.uk/vol1/ftp/}
}
\usage{
kgpe
}
\description{
Sample, pedigree, and population data for 3,202 samples in the expanded 1000 Genomes Project data.
}
\references{
Byrska-Bishop, Marta, et al. "High-coverage whole-genome sequencing of the expanded 1000 Genomes Project cohort including 602 trios." Cell 185.18 (2022): 3426-3440.

1000 Genomes Project Consortium. "A global reference for human genetic variation." Nature 526.7571 (2015): 68.

License information is available at \url{https://github.com/igsr/1000Genomes_data_indexes/blob/master/LICENSE}. The 1000 Genomes data is made publicly available according to the Fort Lauderdale Agreement (\url{https://www.genome.gov/Pages/Research/WellcomeReport0303.pdf}).
}
\keyword{datasets}
