% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp-package.R
\docType{package}
\name{kgp-package}
\alias{kgp}
\alias{kgp-package}
\title{kgp: 1000 Genomes Project Metadata}
\description{
Metadata for samples collected and sequenced for the 1000 Genomes Project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/kgp}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
