% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgp-package.R
\docType{package}
\name{kgp-package}
\alias{kgp}
\alias{kgp-package}
\title{kgp: 1000 Genomes Project Metadata}
\description{
Metadata about populations and data about samples from the 1000 Genomes Project, including the 2,504 samples sequenced for the Phase 3 release and the expanded collection of 3,202 samples with 602 additional trios. The data is described in Auton et al. (2015) \doi{10.1038/nature15393} and Byrska-Bishop et al. (2022) \doi{10.1016/j.cell.2022.08.004}, and raw data is available at \url{http://ftp.1000genomes.ebi.ac.uk/vol1/ftp/}. See Turner (2022) \doi{10.48550/arXiv.2210.00539} for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/kgp}
  \item \url{https://stephenturner.github.io/kgp/}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
